+++ domain/src/main/java/com/inditex/core/platform/domain/model/PriceId.java
@@ -2,8 +2,8 @@ package com.inditex.core.platform.domain.model;
 
 import com.inditex.core.platform.domain.BaseId;
 
-public class PriceId extends BaseId<Long> {
-    public PriceId(Long value) {
+public class PriceId extends BaseId<Integer> {
+    public PriceId(Integer value) {
         super(value);
     }
 }
